+++ qt/python/mantidqt/widgets/codeeditor/execution.py
@@ -20,7 +20,7 @@ from six import PY2, iteritems
 
 # local imports
 from mantidqt.widgets.codeeditor.inputsplitter import InputSplitter
-from mantidqt.utils.async import AsyncTask
+from mantidqt.utils.asynchronous import AsyncTask
 
 if PY2:
     from inspect import getargspec as getfullargspec
